+++ hardware/plarail_detection/hallsensor-hachioji/Hallsensor.hpp
