+++ hardware/plarail_detection/hallsensor-hashimoto/Hallsensor.hpp
