+++ hardware/plarail_detection/hallsensor-sakurajosui/Hallsensor.hpp
