+++ hardware/plarail_detection/hallsensor-shinjuku/Hallsensor.hpp
