+++ hardware/plarail_detection/hallsensor-testver-mul/Hallsensor.hpp
