+++ hardware/plarail_detection/hallsensor-testver/Hallsensor.hpp
